+++ src/test/java/org/jenkinsci/plugins/pipeline/utility/steps/fs/TeeStepTest.java
@@ -39,6 +39,9 @@ import org.jvnet.hudson.test.BuildWatcher;
 import org.jvnet.hudson.test.Issue;
 import org.jvnet.hudson.test.RestartableJenkinsRule;
 
+import static org.hamcrest.Matchers.stringContainsInOrder;
+import static org.junit.Assert.assertThat;
+
 public class TeeStepTest {
 
     @ClassRule
@@ -73,7 +76,8 @@ public class TeeStepTest {
                 SemaphoreStep.success("wait/1", null);
                 WorkflowRun b = r.jenkins.getItemByFullName("p", WorkflowJob.class).getBuildByNumber(1);
                 r.waitForCompletion(b);
-                r.assertLogContains("got: first message second message " + (Functions.isWindows() ? "WS>rem" : "+ true"), b);
+                assertThat(r.getLog(b), stringContainsInOrder("got: first message second message", Functions.isWindows() ? "WS>rem" : "+ true"));
+
         });
     }